# openapi.v2.yaml
openapi: 3.0.3
info:
  title: LODA API v2
  version: "2.0.0"
  description: |
    API for managing, evaluating and submitting LODA programs and integer sequences.
servers:
- url: https://api.loda-lang.org/v2
...
paths:
...
  /submissions:
    get:
      tags: [ Submissions ]
      summary: Get submissions
      description: |
        Retrieve a paginated list of submissions. This endpoint returns all submissions including programs, sequence refresh requests, and b-file removal requests with support for pagination and filtering.
        Submissions contain the ID, submitter, content, submission type (add/update/remove/refresh), and object type (program/sequence/bfile).
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Skip'
      - name: mode
        in: query
        required: false
        description: Filter submissions by mode (add, update, remove, or refresh)
        schema:
          type: string
          enum: [add, update, remove, refresh]
      - name: type
        in: query
        required: false
        description: Filter submissions by object type (program, sequence, or bfile)
        schema:
          type: string
          enum: [program, sequence, bfile]
      - name: submitter
        in: query
        required: false
        description: Filter submissions by submitter name
        schema:
          type: string
      responses:
        '200':
          description: List of submissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionsResult'
    post:
      tags: [ Submissions ]
      summary: Submit a new program, sequence refresh request, or b-file removal/refresh request
      description: |
        Submit a new program, sequence refresh request, or b-file removal/refresh request.
        The request body should contain a JSON object with the submission details including ID, submitter, content, mode, and type.
        
        Submission modes:
        - **add**: Add a new program
        - **update**: Update an existing program
        - **remove**: Remove a program or b-file
        - **refresh**: Trigger a sequence data refresh (for sequence type) or remove a b-file (for bfile type)
        
        Object types:
        - **program**: LODA program submission (supports add, update, and remove modes only - refresh is not allowed)
        - **sequence**: Integer sequence refresh request (only 'refresh' mode is allowed). This triggers an update of the sequence data from the OEIS website and deletes the existing b-file.
        - **bfile**: OEIS b-file removal/refresh request (only 'remove' and 'refresh' modes are allowed). Both modes remove the b-file. When a b-file is removed, it is protected for 24 hours during which it cannot be removed again.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Submission'
            examples:
              programSubmission:
                summary: Program submission example
                value:
                  id: A000045
                  submitter: Alice
                  content: |
                    mov $2,1
                    lpb $0
                      sub $0,2
                      add $2,$1
                      add $1,$2
                    lpe
                    mul $0,$2
                    add $0,$1
                  mode: add
                  type: program
              sequenceRefresh:
                summary: Sequence refresh example
                value:
                  id: A000045
                  submitter: Bob
                  mode: refresh
                  type: sequence
              bfileRefresh:
                summary: B-file refresh/removal example
                value:
                  id: A000045
                  submitter: Charlie
                  mode: refresh
                  type: bfile
      responses:
        '200':
          description: Submission result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResult'
        '400':
          description: Bad request, invalid submission format
...
components:
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of results to return (pagination limit)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    Skip:
      name: skip
      in: query
      required: false
      description: Number of items to skip before starting to collect the result set (pagination offset)
      schema:
        type: integer
        minimum: 0
        default: 0
...
  schemas:
...
    OperationResult:
      type: object
      properties:
        status:
          type: string
          enum: [ success, error ]
          description: Status of the operation
        message:
          type: string
          description: Detailed status or error message
...
    Submission:
      type: object
      required:
        - id
        - mode
        - type
      properties:
        id:
          type: string
          pattern: '^[A-Z]\d{1,10}$'
          description: ID of the program, sequence, or b-file
          example: A000045
        mode:
          type: string
          enum: [add, update, remove, refresh]
          description: |
            Type of submission operation. 
            - Programs support: add, update, remove (refresh is NOT allowed)
            - Sequences support: refresh (only)
            - B-files support: remove, refresh (both behave the same - they remove the b-file)
          example: add
        type:
          type: string
          enum: [program, sequence, bfile]
          description: |
            Type of object being submitted.
            - **program**: LODA program submission (refresh mode not allowed)
            - **sequence**: Sequence refresh request (only 'refresh' mode allowed)
            - **bfile**: OEIS b-file removal/refresh request (only 'remove' and 'refresh' modes allowed)
          example: program
        content:
          type: string
          description: Content of the submission - program code (not required when mode is 'remove' or 'refresh', or type is 'bfile')
          example: |
            mov $2,1
            lpb $0
              sub $0,2
              add $2,$1
              add $1,$2
            lpe
            mul $0,$2
            add $0,$1
        submitter:
          type: string
          description: Name of the person submitting
          example: Alice
    SubmissionsResult:
      type: object
      properties:
        session:
          type: integer
          format: int64
          description: Unix timestamp of the current session
        total:
          type: integer
          description: Total number of submissions
        results:
          type: array
          items:
            $ref: '#/components/schemas/Submission'
...